+++ ecds-bill-invocing/src/main/java/com/bosssoft/ecds/entity/po/UneCbillItem.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -39,12 +40,12 @@ public class UneCbillItem {
      */
     private int fNumber;
 
-    private double fStd;
+    private BigDecimal fStd;
 
     /**
      * 金额
      */
-    private double fAmt;
+    private BigDecimal fAmt;
 
     private int fVersion;
